Correct benchmark doxygen so it appears in documentation
The benchmark documentation was written with the intent that it would be
included in the doxygen render module. It wasn't for two reasons:

1. It used /* */ comments (so was ignored).
2. It was in the .cc file (instead of .h).
